pyproject.toml
[tool.poetry]
name = "pino"
version = "0.1.0"
description = "A python-arduino interface"
authors = ["RidiculousLoop"]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.4"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-language-server = {extras = ["all"], version = "^0.31.9"}
isort = "^4.3.21"
pyflakes = "^2.1.1"
pycodestyle = "^2.5.0"
yapf = "^0.29.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
